+++ ChangeLog.md
@@ -39,6 +39,8 @@
   of a player's positions. (#7)
 - The extended `cwevent` fields reporting runner position correct report pinch-runners as
   position 12 instead of position 0.  (#11)
+- Added command-line switch `-D` to allow referring to TEAM and roster files in a directory
+  different from the current one.  (#51)
 - Support for the use of "?" as a "missing fielder" placeholder has been removed; this has been
   deprecated by Retrosheet for a very long time.  Play strings are normalised to strip out (and
   therefore simply ignore) the presence of any stray "?" characters, to accommodate a handful of